+++ app/Providers/RepositoryServiceProvider.php
@@ -15,6 +15,16 @@ class RepositoryServiceProvider extends ServiceProvider
     public function register()
     {
         $this->app->bind(
+            'App\Interfaces\CloudflareInterface',
+            'App\Repositories\CloudflareRepository'
+        );
+
+        $this->app->bind(
+            'App\Interfaces\Settings\SettingsCloudflareInterface',
+            'App\Repositories\Settings\SettingsCloudflareRepository'
+        );
+
+        $this->app->bind(
             'App\Interfaces\Settings\SettingsUserInterface',
             'App\Repositories\Settings\SettingsUserRepository'
         );