+++ database/factories/SettingsCloudflareFactory.php
@@ -1,7 +1,6 @@
 <?php
 
-use App\User;
-use Illuminate\Support\Str;
+use \App\Repositories\Settings\SettingsCloudflareRepository;
 use Faker\Generator as Faker;
 
 /*
@@ -15,14 +14,10 @@ use Faker\Generator as Faker;
 |
 */
 
-$factory->define(User::class, function (Faker $faker) {
+$factory->define(SettingsCloudflareRepository::class, function (Faker $faker) {
     return [
-        'name' => null,
-        'email' => null,
-        'email_verified_at' => null,
-        'username' => 'domotic',
-        'password' => '$2y$10$92IXUNpkjO0rOQ5byMi.Ye4oKoEa3Ro9llC/.og/at2.uheWG/igi', // password,
-        'last_password_change' => null,
-        'updated_at' => null,
+        'api_key' => $faker->asciify('**************'),
+        'email' => $faker->email(),
+        'domain_list' => $faker->domainName()
     ];
 });