+++ resources/lang/it/validation.php
@@ -2,4 +2,5 @@
 
 return [
     'confirmed' => 'conferma :attribute non corrisponde al precedente',
+    'required' => ':attribute è obbligatorio'
 ];